Move a couple more checks to common
